Moved GameViewActor to "controller" package
