+++ client/src/main/scala/it/cwmp/client/controller/GameViewActor.scala
@@ -1,17 +1,17 @@
-package it.cwmp.client.view.game
+package it.cwmp.client.controller
 
 import java.util.concurrent.ThreadLocalRandom
 
 import akka.actor.{Actor, ActorRef, Cancellable}
+import it.cwmp.client.controller.GameViewActor._
 import it.cwmp.client.controller.game.GameConstants.{MAX_TIME_BETWEEN_CLIENT_SYNCHRONIZATION, MIN_TIME_BETWEEN_CLIENT_SYNCHRONIZATION}
 import it.cwmp.client.controller.game.GameEngine
 import it.cwmp.client.controller.messages.Initialize
-import it.cwmp.client.controller.{ActorAlertManagement, AlertMessages}
 import it.cwmp.client.model.game.distributed.AkkaDistributedState.UpdateState
 import it.cwmp.client.model.game.impl._
 import it.cwmp.client.utils.GeometricUtils
-import it.cwmp.client.view.game.GameViewActor._
 import it.cwmp.client.view.game.drawing.GameFX
+import it.cwmp.client.view.game.{CellView, TentacleView}
 import it.cwmp.client.view.{FXAlertsController, FXRunOnUIThread}
 import it.cwmp.utils.Logging
 import it.cwmp.utils.Utils.stringToOption
